Wow big stuff (logic, style, animations)
- Logic: DONE.
Note that the picker behaves like this:
If the key pressed is greater than 2 (for hours) and 5 (for minutes) it
assumes it's a unit and precede it with a 0.
Example: if you select 3 from hours it will display 03.
The other solution would be to explicitly ask for the leading 0 when
selecting the first digit by disabling the keys greater than 2 (for
hours) and 5 (for minutes)
- styled perfected (colors and fonts)
- animations on the display to show the currently selected unit (hours
or minutes)
- added a clear button to reset the display to Now